cleanup and move pool stuff into own source files
