+++ postgres/CMakeLists.txt
@@ -18,6 +18,8 @@ set(SOURCE_FILES
     src/main/c/am/zdbseqscan.h
     src/main/c/am/zdbscore.c
     src/main/c/am/zdbscore.h
+    src/main/c/am/zdbpool.c
+    src/main/c/am/zdbpool.h
     src/main/c/rest/rest.c
     src/main/c/rest/rest.h
     src/main/c/util/curl_support.c
