+++ postgres/src/main/c/am/zdb_interface.c
@@ -27,7 +27,7 @@
 #include "utils/memutils.h"
 
 #include "zdb_interface.h"
-#include "util/zdbutils.h"
+#include "zdbpool.h"
 #include "elasticsearch.h"
 
 relopt_kind RELOPT_KIND_ZDB;
@@ -141,10 +141,11 @@ ZDBIndexDescriptor *zdb_alloc_index_descriptor(Relation indexRel)
 	desc->tableName      = pstrdup(RelationGetRelationName(heapRel));
 	desc->options		 = ZDBIndexOptionsGetOptions(indexRel) == NULL ? NULL : pstrdup(ZDBIndexOptionsGetOptions(indexRel));
 
-	desc->searchPreference = ZDBIndexOptionsGetSearchPreference(indexRel);
-	desc->bulk_concurrency = ZDBIndexOptionsGetBulkConcurrency(indexRel);
-	desc->batch_size       = ZDBIndexOptionsGetBatchSize(indexRel);
-	desc->fieldLists       = ZDBIndexOptionsGetFieldLists(indexRel);
+	desc->searchPreference   = ZDBIndexOptionsGetSearchPreference(indexRel);
+	desc->bulk_concurrency   = ZDBIndexOptionsGetBulkConcurrency(indexRel);
+	desc->batch_size         = ZDBIndexOptionsGetBatchSize(indexRel);
+	desc->fieldLists         = ZDBIndexOptionsGetFieldLists(indexRel);
+	desc->current_pool_index = InvalidPoolIndex;
 
 	if (desc->isShadow)
 	{
